# fill_via_chunks: use safe code via chunks_exact_mut on BE