@Ralith 2021-09-11
Nice, that worked out better than I expected! It's interesting to see what LLVM is and isn't capable of optimizing out here.