@dhardy 2021-09-12
> One approach I've seen before for endianness handling is to just do the copy and then go back and do a second pass afterwards to fix the endianness in-place

Doing this is problematic since we know nothing about the alignment of the output byte-array, and may have a partial word at the end. But doing the reverse works fine, and there doesn't appear to be a reason we can't mutate the source (though it is a breaking change). These are the results I get (byte-flipping to BE on my machine, since I don't have a BE test machine):

```
test gen_bytes_chacha12      ... bench:     273,486 ns/iter (+/- 531) = 3744 MB/s
test gen_bytes_chacha20      ... bench:     411,202 ns/iter (+/- 1,400) = 2490 MB/s
test gen_bytes_chacha8       ... bench:     203,651 ns/iter (+/- 953) = 5028 MB/s
```