@dhardy 2021-09-12
Regarding Ben Sanders comment:

> This is sound if we assume src and dst are both slices of with elements that have no invalid bitpatterns. However, that's not statically checked, and this macro can be invoked in safe code, so it's a bit sketchy. It would be better as an unsafe generic function, though the comment suggests that we'd want to be careful to ensure it gets inlined.

My take is that (1) the macro is private to the module and (2) that the only way to make it a generic function would be to define our own trait (or use `num-traits`), since `to_le` is not covered by any `std` trait. Alternatives are (a) removing all `unsafe`, (b) inlining or (c) pushing the `unsafe` up to the function using the macro (thus making all of the macro code `unsafe`). My current preference is just to leave this as it is, but I'm open to further discussion.