@joshlf 2021-09-12
> Doing this is problematic since we know nothing about the alignment of the output byte-array, and may have a partial word at the end.

I believe you're guaranteed not to have a partial word at the end because the source's length is a multiple of the chunk size.

As for the alignment, I suppose I could see that being a performance issue, but it's probably worth benchmarking so you don't have to break the API. You obviously have a better sense than I do of how big of a deal that kind of breaking change is, though.