@Ralith 2021-09-12
> (c) pushing the unsafe up to the function using the macro (thus making all of the macro code unsafe). My current preference is just to leave this as it is, but I'm open to further discussion.

Ben here. I think (c) is a large improvement because it makes it clear at the macro callsite that statically unchecked soundness requirements are in play. The current form makes it easy to miss that for a reader who's not already familiar with the code. Nothing wrong with unsafe private helpers, but it's a lot clearer when that unsafety isn't hidden.

I'm a fan of replacing macros with traits whenever possible since macros are a much bigger hammer, but I understand if you'd prefer to avoid the boilerplate.