@dhardy 2021-09-13
> I believe you're guaranteed not to have a partial word at the end because the source's length is a multiple of the chunk size.

No it's not. It often will be in practice but the destination is just whatever byte-slice was passed to `rng.fill_bytes`.

I *think* the API breakage is fine, though obviously it delays releasing this fix. We could split this PR and do that part later, I guess (it's only a perf boost). @vks @newpavlov?

Thanks for the arguments Ben. Traits and generics are also a pretty big hammer, but still it doesn't have a noticeable compile-time impact. Your argument:

> This is sound if we assume src and dst are both slices of with elements that have no invalid bitpatterns

.. still partly applies with a generic function: now we know `dest: &mut [u8]`, so *writing* any bit-pattern is valid, but `src: &mut [T]` implies we don't know what we're reading (and we don't have `bytemuck::Pod` to add as a trait bound). But I think this is good enough that local usage of `unsafe` is fine?