@vks 2021-09-13
@dhardy 
> .. still partly applies with a generic function: now we know `dest: &mut [u8]`, so _writing_ any bit-pattern is valid, but `src: &mut [T]` implies we don't know what we're reading (and we don't have `bytemuck::Pod` to add as a trait bound). But I think this is good enough that local usage of `unsafe` is fine?

Shouldn't `ToLe` be an unsafe trait then?