@dhardy 2021-09-13
> Shouldn't ToLe be an unsafe trait then?

But is there anything `unsafe` about performing a byte copy *from* the type? Sure, forgetting to run a destructor can break some things, but it can't break *memory safety*.

Anyway, I think I should move the breaking parts to another PR so that we can do another patch release.