@dhardy 2021-09-13
I removed the breaking changes. The results are actually a little better than before, though I don't understand why:
```
test gen_bytes_chacha12      ... bench:     302,031 ns/iter (+/- 2,346) = 3390 MB/s
test gen_bytes_chacha20      ... bench:     437,592 ns/iter (+/- 4,405) = 2340 MB/s
test gen_bytes_chacha8       ... bench:     230,950 ns/iter (+/- 1,861) = 4433 MB/s
```