@dhardy 2021-09-13
> `src: &mut [T]` may contain invalid byte patterns

What's an invalid byte pattern to `[u8]`?

---

Your code is marginally slower than mine.