@joshlf 2021-09-13
> It often will be in practice but the destination is just whatever byte-slice was passed to `rng.fill_bytes`.

Ack.