@vks 2021-09-13
@dhardy 
> What's an invalid byte pattern to [u8]?

Fair enough, I can only think of the undefined behavior being introduced by an unsound implementation of `ToLe`.

> Your code is marginally slower than mine.

Thanks for checking! How much slower? Is it fast enough so we can use the safe code?