@vks 2021-09-13
Looks like value stability on big endian platforms was broken:

```
---- rngs::std::test::test_stdrng_construction stdout ----
thread 'main' panicked at 'assertion failed: `(left == right)`
  left: `[10719222850664546238, 12020057239787804515]`,
 right: `[10719222850664546238, 14064965282130556830]`', src/rngs/std.rs:96:9
```