@Ralith 2021-09-13
In the current form, I believe either `fill_via_chunks` or (preferably) `trait ToLe` should, strictly speaking, be marked unsafe, because they allow the observation of uninitialized padding bytes in a hypothetical padded `T` on little-endian targets.