@dhardy 2021-09-14
> Thanks for checking! How much slower? Is it fast enough so we can use the safe code?

Sorry, I meant it's a couple of percent slower than my "safe" code (which performs as below on LE):
```
test gen_bytes_chacha12      ... bench:     290,558 ns/iter (+/- 2,151) = 3524 MB/s
test gen_bytes_chacha20      ... bench:     427,874 ns/iter (+/- 7,506) = 2393 MB/s
test gen_bytes_chacha8       ... bench:     219,347 ns/iter (+/- 2,568) = 4668 MB/s
```

Good point @Ralith. That implies *both* the function and the trait are unsafe.